Removing type="text/javascript" requirement
Removing the requirement that scripts have the type="text/javascript" attribute in order to be executed.  The current HTML spec states that <script type=""> should be omitted because it is redundant.  Since this is the expectation for many developers, I'm suggesting that we remove this requirement from HTMX.

SOURCE: https://www.w3.org/TR/html53/semantics-scripting.html#element-attrdef-script-type

https://stackoverflow.com/questions/4243577
https://stackoverflow.com/questions/13883553